[WFLY-11270]: OpenTracing tests are in legacy org.jboss.as package in widlfly-ts-integ-basic
* Moving the tests to the proper package

Jira: https://issues.jboss.org/browse/WFLY-11270
