+++ testsuite/integration/basic/src/test/java/org/wildfly/test/integration/microprofile/opentracing/ResourceNotTracedTestCase.java
@@ -1,4 +1,4 @@
-package org.jboss.as.test.integration.microprofile.opentracing;
+package org.wildfly.test.integration.microprofile.opentracing;
 
 import static org.jboss.as.test.shared.integration.ejb.security.PermissionUtils.createPermissionsXmlAsset;
 
@@ -9,10 +9,9 @@ import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
 import org.jboss.as.test.integration.common.HttpRequest;
-import org.jboss.as.test.integration.microprofile.opentracing.application.MockTracerFactory;
-import org.jboss.as.test.integration.microprofile.opentracing.application.OpenTracingApplication;
-import org.jboss.as.test.integration.microprofile.opentracing.application.TracedBean;
-import org.jboss.as.test.integration.microprofile.opentracing.application.WithBeanTracedEndpoint;
+import org.wildfly.test.integration.microprofile.opentracing.application.MockTracerFactory;
+import org.wildfly.test.integration.microprofile.opentracing.application.NotTracedEndpoint;
+import org.wildfly.test.integration.microprofile.opentracing.application.OpenTracingApplication;
 import org.jboss.as.test.shared.TestSuiteEnvironment;
 import org.jboss.shrinkwrap.api.Archive;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
@@ -23,12 +22,12 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import javax.inject.Inject;
-import java.net.SocketPermission;
 import java.net.URL;
+import java.net.SocketPermission;
 import java.util.concurrent.TimeUnit;
 
 @RunWith(Arquillian.class)
-public class ResourceWithCDITestCase {
+public class ResourceNotTracedTestCase {
     @Inject
     Tracer tracer;
 
@@ -38,7 +37,7 @@ public class ResourceWithCDITestCase {
     @Deployment
     public static Archive<?> deploy() {
         WebArchive war = ShrinkWrap.create(WebArchive.class);
-        war.addClass(ResourceTracedTestCase.class);
+        war.addClass(ResourceNotTracedTestCase.class);
 
         war.addClass(MockTracerFactory.class);
         war.addPackage(MockTracer.class.getPackage());
@@ -47,8 +46,7 @@ public class ResourceWithCDITestCase {
         war.addAsWebInfResource(EmptyAsset.INSTANCE, "beans.xml");
 
         war.addClass(OpenTracingApplication.class);
-        war.addClass(WithBeanTracedEndpoint.class);
-        war.addClass(TracedBean.class);
+        war.addClass(NotTracedEndpoint.class);
 
         war.addClass(HttpRequest.class);
 
@@ -63,16 +61,17 @@ public class ResourceWithCDITestCase {
     }
 
     @Test
-    public void tracedEndpointYieldsSpan() throws Exception {
+    public void notTracedEndpointYieldsNoSpans() throws Exception {
         Assert.assertTrue(tracer instanceof MockTracer);
         MockTracer mockTracer = (MockTracer) tracer;
 
-        performCall("opentracing/with-bean");
+        performCall("opentracing/not-traced");
 
-        Assert.assertEquals(3, mockTracer.finishedSpans().size());
+        Assert.assertEquals(0, mockTracer.finishedSpans().size());
     }
 
     private void performCall(String path) throws Exception {
         HttpRequest.get(url + path, 10, TimeUnit.SECONDS);
     }
+
 }
