+++ testsuite/integration/basic/src/test/java/org/wildfly/test/integration/microprofile/opentracing/application/OpenTracingApplication.java
@@ -1,4 +1,4 @@
-package org.jboss.as.test.integration.microprofile.opentracing.application;
+package org.wildfly.test.integration.microprofile.opentracing.application;
 
 import javax.ws.rs.ApplicationPath;
 import javax.ws.rs.core.Application;
